Update .config/nvim/lua/autocmds.lua Update .config/nvim/lua/options.lua Update .config/nvim/lua/plugins/obsidian.lua Update .config/nvim/spell/en.utf-8.add Update .config/nvim/spell/en.utf-8.add.spl Update .config/nvim/spell/medical.utf-8

+++ dot_config/nvim/lua/autocmds.lua
@@ -47,13 +47,23 @@ vim.api.nvim_create_autocmd("BufEnter", {
 	desc = "Remove colorcolumn in nofile buffers",
 })
 
--- Chezmoi
-local chezmoi = vim.api.nvim_create_augroup("Chezmoi", { clear = true })
+-- Syncing Config with Windows and chezmoi
+local confsync = vim.api.nvim_create_augroup("ConfigSync", { clear = true })
+
 vim.api.nvim_create_autocmd("BufWritePost", {
-	group = chezmoi,
+	group = confsync,
 	pattern = "*/.local/share/chezmoi/*",
 	callback = function()
-		vim.system({ "chezmoi", "apply" }, {
+		local wsl = vim.fn.stdpath("config")
+		local win = vim.fn.expand("$HOME/windows/AppData/Local/nvim")
+		local ch = vim.fn.expand("$HOME/.local/share/chezmoi/dot_config/nvim")
+
+		if vim.fn.isdirectory(win) then
+			-- add new spellings from windows before overwriting everything
+			vim.system({ "rsync", "-rtu", win .. "/spell/", ch .. "/spell" }, { text = true }, function(_) end)
+		end
+
+		local apply = vim.system({ "chezmoi", "apply" }, {
 			stdout = function(_, data)
 				if data ~= nil then
 					vim.notify("Apply failed.\nRun chezmoi apply from command line.", vim.log.levels.ERROR)
@@ -66,34 +76,21 @@ vim.api.nvim_create_autocmd("BufWritePost", {
 			end,
 			text = true,
 		}, function(_) end)
-	end,
-	desc = "Apply changes in chezmoi files to local files (chezmoi source dir -> destination)",
-})
 
--- Syncing Config with Windows
-local winsync = vim.api.nvim_create_augroup("WindowsSync", { clear = true })
-
-vim.api.nvim_create_autocmd("BufWritePost", {
-	group = winsync,
-	pattern = "*/.local/share/chezmoi/*",
-	callback = function()
-		local wsl = vim.fn.stdpath("config")
-		local win = vim.fn.expand("$HOME/windows/AppData/Local/nvim")
-
-		-- add new spellings from windows before overwriting everything
-		vim.system({ "rsync", "-rt", win .. "/spell/", wsl .. "spell" }, { text = true }, function(_) end)
-
-		local cmd = { "rsync", "-a", "--delete", "--exclude", ".git", "--exclude", "lazy-lock.json", wsl .. "/", win }
-		vim.system(cmd, {
-			text = true,
-		}, function(completed)
-			local code = completed.code
-			if code == 0 then
-				vim.notify("Config synced to Windows", vim.log.levels.INFO)
-			else
-				vim.notify("rsync failed (code " .. code .. ")", vim.log.levels.ERROR)
-			end
-		end)
+		if vim.fn.isdirectory(win) and apply:wait().code == 0 then
+			local cmd =
+				{ "rsync", "-a", "--delete", "--exclude", ".git", "--exclude", "lazy-lock.json", wsl .. "/", win }
+			vim.system(cmd, {
+				text = true,
+			}, function(completed)
+				local code = completed.code
+				if code == 0 then
+					vim.notify("Config synced to Windows", vim.log.levels.INFO)
+				else
+					vim.notify("rsync failed (code " .. code .. ")", vim.log.levels.ERROR)
+				end
+			end)
+		end
 	end,
 	desc = "Push edited config file to Windows via rsync",
 })
